api/requirements.txt
fastapi
uvicorn
pymongo
motor
python-multipart
python-dotenv
python-dateutil